Split up Node.js automation tests and run with `--parallel`
Moch can parallelize test runs by file. Split up `sdk/nodejs/tests/automation/localWorkspace.spec.ts` to enable this.

Also fixes some cases where we were missing an `await` on `assert.rejects`, and some version test oddness that was uncovered after fixing the awaits.
