+++ sdk/nodejs/tests/automation/cmd.spec.ts
@@ -73,9 +73,9 @@ describe("automation/cmd", () => {
         it("throws if the found version is not compatible with the requested version", async () => {
             const installedVersion = new semver.SemVer("3.97.0");
             await PulumiCommand.install({ version: installedVersion });
-            const requestedVersion = new semver.SemVer("3.98.0");
-            assert.rejects(PulumiCommand.get({ version: requestedVersion }));
-            assert.doesNotReject(PulumiCommand.get({ version: installedVersion, skipVersionCheck: true }));
+            const requestedVersion = new semver.SemVer("10.0.0");
+            await assert.rejects(PulumiCommand.get({ version: requestedVersion }));
+            await assert.doesNotReject(PulumiCommand.get({ version: installedVersion, skipVersionCheck: true }));
         });
     });
