+++ sdk/nodejs/tests/automation/util.ts
@@ -13,6 +13,9 @@
 // limitations under the License.
 
 import { randomUUID } from "crypto";
+import * as tmp from "tmp";
+
+import { LocalWorkspaceOptions, ProjectRuntime } from "../../automation";
 
 /** @internal */
 export function getTestSuffix() {
@@ -28,4 +31,110 @@ export function getTestOrg() {
         return "pulumi-test";
     }
     return "organization";
+}
+
+/**
+ * Augments the provided {@link LocalWorkspaceOptions} so that they reference a
+ * either a file backend or a cloud backend, depending on whether PULUMI_ACCESS_TOKEN
+ * is set in the environment.
+ *
+ * @internal
+ */
+export function withTestBackend(
+    opts?: LocalWorkspaceOptions,
+    name?: string,
+    description?: string,
+    runtime?: string,
+): LocalWorkspaceOptions {
+    if (process.env.PULUMI_ACCESS_TOKEN) {
+        return withCloudBackend(opts, name, description, runtime);
+    }
+    return withTemporaryFileBackend(opts, name, description, runtime);
+}
+
+function withCloudBackend(
+    opts?: LocalWorkspaceOptions,
+    name?: string,
+    description?: string,
+    runtime?: string,
+): LocalWorkspaceOptions {
+    let url = "https://api.pulumi.com";
+    if (process.env.PULUMI_BACKEND_URL) {
+        url = process.env.PULUMI_BACKEND_URL;
+    }
+    const backend = {
+        url: url,
+    };
+    if (name === undefined) {
+        name = "node_test";
+    }
+    if (runtime === undefined) {
+        runtime = "nodejs";
+    }
+    return {
+        ...opts,
+        projectSettings: {
+            // We are obliged to provide a name and runtime if we provide project
+            // settings, so we do so, but we spread in the provided project settings
+            // afterwards so that the caller can override them if need be.
+            name: name,
+            runtime: runtime as ProjectRuntime,
+            description: description,
+
+            ...opts?.projectSettings,
+            backend,
+        },
+    };
+}
+
+function withTemporaryFileBackend(
+    opts?: LocalWorkspaceOptions,
+    name?: string,
+    description?: string,
+    runtime?: string,
+): LocalWorkspaceOptions {
+    const tmpDir = tmp.dirSync({
+        prefix: "nodejs-tests-automation-",
+        unsafeCleanup: true,
+    });
+
+    const backend = { url: `file://${tmpDir.name}` };
+
+    if (name === undefined) {
+        name = "node_test";
+    }
+    if (runtime === undefined) {
+        runtime = "nodejs";
+    }
+
+    return withTestConfigPassphrase({
+        ...opts,
+        pulumiHome: tmpDir.name,
+        projectSettings: {
+            // We are obliged to provide a name and runtime if we provide project
+            // settings, so we do so, but we spread in the provided project settings
+            // afterwards so that the caller can override them if need be.
+            name: name,
+            runtime: runtime as ProjectRuntime,
+            description: description,
+
+            ...opts?.projectSettings,
+            backend,
+        },
+    });
+}
+
+/**
+ * Augments the provided {@link LocalWorkspaceOptions} so that they set up an
+ * environment containing a test `PULUMI_CONFIG_PASSPHRASE` variable suitable
+ * for use with a local file backend.
+ */
+function withTestConfigPassphrase(opts?: LocalWorkspaceOptions): LocalWorkspaceOptions {
+    return {
+        ...opts,
+        envVars: {
+            ...opts?.envVars,
+            PULUMI_CONFIG_PASSPHRASE: "test",
+        },
+    };
 }
